Add capability to run k8s-image-puller locally
Reorganize some of the code to allow running a local version of
kubernetes image puller, for testing.
